+++ src/modules/service/src/main/java/com/opentravelsoft/service/resource/AirwaysService.java
@@ -14,7 +14,7 @@ public interface AirwaysService {
 
   public Airways roGetAirwaysDetail(String airwaysId);
 
-  public PaginationSupport getAirwaysList(boolean isActive, long fromRecord,
+  public PaginationSupport getAirwaysList(boolean isActive, int fromRecord,
       int moveCount);
 
   List<Airways> getAirwaysList(boolean isActive);
